+++ packages/jest-core/src/TestScheduler.ts
@@ -347,7 +347,10 @@ class TestScheduler {
             : this.addReporter(new DefaultReporter(this._globalConfig));
           break;
         case 'github-actions':
-          GITHUB_ACTIONS && this.addReporter(new GitHubActionsReporter());
+          GITHUB_ACTIONS &&
+            this.addReporter(
+              new GitHubActionsReporter(this._globalConfig, options),
+            );
           break;
         case 'summary':
           summary = true;