+++ packages/jest-types/src/Circus.ts
@@ -185,6 +185,7 @@ export type TestResult = {
   invocations: number;
   status: TestStatus;
   location?: {column: number; line: number} | null;
+  numPassingAsserts: number;
   retryReasons: Array<FormattedError>;
   testPath: Array<TestName | BlockName>;
 };
@@ -245,6 +246,7 @@ export type TestEntry = {
   mode: TestMode;
   concurrent: boolean;
   name: TestName;
+  numPassingAsserts: number;
   parent: DescribeBlock;
   startedAt?: number | null;
   duration?: number | null;